Implement the Encoder trait for PacketCodec
Take a StatusCode and generate the corresponding packet with alll
required fields. Write out the response packet to the provided
buffer.

Remove integer values from the StatusCode enum and instead let the encoder
parse a StatusCode and write a status code integer value.

Write unit tests for the packet encoder.